+++ dutil-basic/src/main/java/com/dwarfeng/dutil/basic/num/unit/Time.java
@@ -11,37 +11,37 @@ import com.dwarfeng.dutil.basic.num.NumberValue;
 public enum Time implements NumberValue {
 
     /**
-     * 纳秒
+     * 纳秒。
      */
     NS(86400000000000d),
 
     /**
-     * 微秒
+     * 微秒。
      */
     US(86400000000d),
 
     /**
-     * 毫秒
+     * 毫秒。
      */
     MS(86400000d),
 
     /**
-     * 秒
+     * 秒。
      */
     SEC(86400d),
 
     /**
-     * 分钟
+     * 分钟。
      */
     MIN(1440d),
 
     /**
-     * 小时
+     * 小时。
      */
     HOR(24d),
 
     /**
-     * 天
+     * 天。
      */
     DAY(1d);